# src/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "version": "1.0.0",
        "title": "Task management",
        "description": "An API to manage my personal tasks.",
        "contact": {
            "name": "Saul Dominguez",
            "url": "https://sauldom.dev/",
            "email": "saulydominguez@gmail.com"
        }
    },
    "paths": {
...
        "/user/": {
            "get": {
                "summary": "Get current user information",
                "tags": ["User"],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "default": "ok"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "summary": "Update current user information",
                "tags": ["User"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserBody"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "default": "ok"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
    "security": [
        {
            "apiKey": []
        }
    ],
    "components": {
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
...
            "UpdateUserBody": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "emailConfirmed": {
                        "type": "boolean"
                    },
                    "username": {
                        "type": "string"
                    },
                    "onboarded": {
                        "type": "boolean"
                    }
                },
                "required": ["id", "email", "emailConfirmed", "onboarded"]
            },
...
        }
    }
}